Conflicts:
	kundera-cassandra/src/main/java/com/impetus/client/cassandra/CassandraClientBase.java
	kundera-cassandra/src/main/java/com/impetus/client/cassandra/config/CassandraPropertyReader.java
	kundera-cassandra/src/main/java/com/impetus/client/cassandra/thrift/ThriftClient.java
	kundera-cassandra/src/main/java/com/impetus/client/cassandra/thrift/ThriftClientFactory.java

+++ kundera-tests/src/test/java/com/impetus/kundera/tests/crossdatastore/useraddress/AssociationBase.java
@@ -48,10 +48,12 @@ import com.impetus.client.redis.RedisPropertyReader;
 import com.impetus.kundera.PersistenceProperties;
 import com.impetus.kundera.client.Client;
 import com.impetus.kundera.metadata.KunderaMetadataManager;
+import com.impetus.kundera.metadata.model.CoreMetadata;
 import com.impetus.kundera.metadata.model.EntityMetadata;
 import com.impetus.kundera.metadata.model.KunderaMetadata;
 import com.impetus.kundera.metadata.model.MetamodelImpl;
 import com.impetus.kundera.metadata.model.PersistenceUnitMetadata;
+import com.impetus.kundera.proxy.cglib.CglibLazyInitializerFactory;
 import com.impetus.kundera.tests.cli.CassandraCli;
 import com.impetus.kundera.tests.cli.CleanupUtilities;
 import com.impetus.kundera.tests.crossdatastore.useraddress.dao.UserAddressDaoImpl;
@@ -91,10 +93,14 @@ public abstract class AssociationBase
 
     protected List<Object> col = new ArrayList<Object>();
 
+//    private String persistenceUnits = "redis,rdbms,addMongo,oracle_kvstore,piccandra,secIdxAddCassandra,picongo,addCassandra";
+//    protected static final String[] ALL_PUs_UNDER_TEST = new String[] { "addMongo", "rdbms", "redis", "addCassandra",
+//            "oracle_kvstore" , "addHbase"};
+
     private String persistenceUnits = "rdbms,redis,addMongo,oracle_kvstore,piccandra,secIdxAddCassandra,picongo,addCassandra";
 
     protected static final String[] ALL_PUs_UNDER_TEST = new String[] { "addMongo", "rdbms", "redis", "addCassandra",
-            "oracle_kvstore" /* , "addHbase" */};
+    "oracle_kvstore" /* , "addHbase" */};
 
     protected RDBMSCli cli;
 
@@ -202,6 +208,10 @@ public abstract class AssociationBase
                 // clazz, mAdd);
                 PersistenceUnitMetadata puMetadata = KunderaMetadata.INSTANCE.getApplicationMetadata()
                         .getPersistenceUnitMetadata(pu);
+                
+                CoreMetadata coreMetadata = new CoreMetadata();
+                coreMetadata.setLazyInitializerFactory(new CglibLazyInitializerFactory());
+                KunderaMetadata.INSTANCE.setCoreMetadata(coreMetadata);
 
                 String client = puMetadata.getProperties().getProperty(PersistenceProperties.KUNDERA_CLIENT_FACTORY);
                 if (client.equalsIgnoreCase("com.impetus.client.cassandra.pelops.PelopsClientFactory")
@@ -362,19 +372,22 @@ public abstract class AssociationBase
     private void truncateRedis()
 
     {
-        PersistenceUnitMetadata puMetadata = KunderaMetadata.INSTANCE.getApplicationMetadata()
-                .getPersistenceUnitMetadata("redis");
-        Properties props = puMetadata.getProperties();
-        String contactNode = RedisPropertyReader.rsmd.getHost() != null ? RedisPropertyReader.rsmd.getHost()
-                : (String) props.get(PersistenceProperties.KUNDERA_NODES);
-        String defaultPort = RedisPropertyReader.rsmd.getPort() != null ? RedisPropertyReader.rsmd.getPort()
-                : (String) props.get(PersistenceProperties.KUNDERA_PORT);
-        String password = RedisPropertyReader.rsmd.getPassword() != null ? RedisPropertyReader.rsmd.getPassword()
-                : (String) props.get(PersistenceProperties.KUNDERA_PASSWORD);
-        Jedis connection = new Jedis(contactNode, Integer.valueOf(defaultPort));
-        connection.auth(password);
-        connection.connect();
-        connection.flushDB();
+        if (RedisPropertyReader.rsmd != null)
+        {
+            PersistenceUnitMetadata puMetadata = KunderaMetadata.INSTANCE.getApplicationMetadata()
+                    .getPersistenceUnitMetadata("redis");
+            Properties props = puMetadata.getProperties();
+            String contactNode = RedisPropertyReader.rsmd.getHost() != null ? RedisPropertyReader.rsmd.getHost()
+                    : (String) props.get(PersistenceProperties.KUNDERA_NODES);
+            String defaultPort = RedisPropertyReader.rsmd.getPort() != null ? RedisPropertyReader.rsmd.getPort()
+                    : (String) props.get(PersistenceProperties.KUNDERA_PORT);
+            String password = RedisPropertyReader.rsmd.getPassword() != null ? RedisPropertyReader.rsmd.getPassword()
+                    : (String) props.get(PersistenceProperties.KUNDERA_PASSWORD);
+            Jedis connection = new Jedis(contactNode, Integer.valueOf(defaultPort));
+            connection.auth(password);
+            connection.connect();
+            connection.flushDB();
+        }
     }
 
     /**
